+++ app/controllers/analyses_controller.rb
@@ -6,13 +6,14 @@ class AnalysesController < ApplicationController
 
   def execute
     attributes = params.permit(*analysis_params)
-    absent_keys = analysis_params - attributes.symbolize_keys.keys
+
+    absent_keys = analysis_params - attributes.keys.map(&:to_sym)
     unless absent_keys.empty?
       error_codes = absent_keys.sort.map {|key| "absent_param_#{key}" }
       raise BadRequest, error_codes
     end
 
-    analysis = Analysis.new(attributes.merge(state: 'processing'))
+    analysis = Analysis.new(attributes.merge(state: Analysis::STATE_PROCESSING))
     unless analysis.save
       error_codes = analysis.errors.messages.keys.sort.map do |key|
         "invalid_param_#{key}"