+++ app/jobs/analysis_job.rb
@@ -1,4 +1,4 @@
-class AnalysisJob < ActiveJob::Base
+class AnalysisJob < ApplicationJob
   queue_as :regulus
 
   def perform(analysis_id)
@@ -12,9 +12,7 @@ class AnalysisJob < ActiveJob::Base
       analysis.pair,
       analysis.batch_size,
     ]
-    command = "sudo docker exec regulus python /opt/scripts/learn.py #{args.join(' ')}"
-    ret = system(command)
-    raise StandardError unless ret
+    execute_script('learn.py', args)
 
     from = File.join(script_dir, 'tmp')
     to = Rails.root.join('tmp', 'models', analysis_id.to_s)
@@ -25,7 +23,9 @@ class AnalysisJob < ActiveJob::Base
     analysis.update!(state: 'completed')
     AnalysisMailer.completed(analysis).deliver_now
     FileUtils.rm_rf("#{Rails.root}/tmp/models/#{analysis_id}")
-  rescue StandardError
+  rescue StandardError => e
+    Rails.logger.error(e.message)
+    Rails.logger.error(e.backtrace.join("\n"))
     analysis.update!(state: 'error')
     AnalysisMailer.error(analysis).deliver_now
   end