+++ config/routes.rb
@@ -4,4 +4,9 @@ Rails.application.routes.draw do
 
   get '/predictions' => 'predictions#manage'
   post '/predictions' => 'predictions#execute'
+  post '/predictions/settings' => 'predictions#settings'
+
+  namespace :api, format: 'json' do
+    resources :predictions, only: %i[index]
+  end
 end