+++ db/migrate/20190713031725_add_means_to_predictions.rb
@@ -1,0 +1,8 @@
+class AddMeansToPredictions < ActiveRecord::Migration[4.2]
+  def change
+    add_column :predictions,
+               :means,
+               :string,
+               null: false, default: 'manual', after: :pair
+  end
+end