+++ db/migrate/20190814104812_add_prediction_id_to_predictions.rb
@@ -1,0 +1,8 @@
+class AddPredictionIdToPredictions < ActiveRecord::Migration[5.0]
+  def change
+    add_column :predictions,
+               :prediction_id,
+               :string,
+               null: false, default: '', after: :id
+  end
+end