+++ spec/factories/query.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :query do
+    means { 'auto' }
+    page {}
+    per_page {}
+  end
+end