+++ spec/support/helpers/prediction_helper.rb
@@ -1,0 +1,9 @@
+# coding: utf-8
+
+module PredictionHelper
+  module_function
+
+  def response_keys
+    @response_keys ||= %w[prediction_id model from to pair means result state created_at]
+  end
+end